🎨 Design : 마이페이지, 설정 페이지 UI 수정사항 반영, Setting Modal 디렉토리 이동
