+++ src/components/ToggleArea/index.tsx
@@ -4,10 +4,10 @@ import * as S from './styles'
 export default function ToggleArea() {
   return (
     <S.ToggleArea>
-      <ToggleBox setting={'friendRequests'} type='lg' />
-      <ToggleBox setting={'familyWalkNotifications'} type='lg' />
       <ToggleBox setting={'myWalkNotifications'} type='lg' />
       <ToggleBox setting={'messages'} type='lg' />
+      {/* <ToggleBox setting={'friendRequests'} type='lg' /> */}
+      {/* <ToggleBox setting={'familyWalkNotifications'} type='lg' /> */}
     </S.ToggleArea>
   )
 }
